requirements.txt
discord.py
pandas
scipy
requests
pyyaml
openai
openpyxl
StrEnum
google-cloud-vision
aiohttp_retry